break AudioMixer slave/stats into own files
